# src/main/java/mobile/order/api/PhoneorderApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package mobile.order.api;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import mobile.order.model.InternalErrorResponse;
import mobile.order.model.Order;
import mobile.order.model.OrderAmount;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-07-07T15:38:07.949Z")

@Api(value = "phoneorder", description = "the phoneorder API")
public interface PhoneorderApi {

    @ApiOperation(value = "Process a phone order", nickname = "buyPhone", notes = "Buy one phone from Catalog. ", response = OrderAmount.class, tags={ "setPhoneOrder", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful response", response = OrderAmount.class),
        @ApiResponse(code = 501, message = "Not found", response = InternalErrorResponse.class) })
    @RequestMapping(value = "/phoneorder/",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<OrderAmount> buyPhone(@ApiParam(value = "Data of client and phones ordered" ,required=true )  @Valid @RequestBody Order body);

}